+++ android/domain/src/main/java/com/digginroom/model/user/JoinState.kt
@@ -1,4 +1,4 @@
-package com.digginroom.model.join
+package com.digginroom.model.user
 
 sealed interface JoinState {
